smart_ward/requirements.txt
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
bidict==0.23.1
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
comtypes==1.4.2
config==0.5.1
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
Django==4.2.11
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-session-timeout==0.1.0
django-socketio==0.3.9
djangorestframework==3.15.1
gevent==24.2.1
gevent-socketio==0.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
msgpack==1.0.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
service-identity==24.1.0
setuptools==69.5.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sphinx-me==0.3
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
utils==1.0.2
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wheel==0.36.2
whitenoise==6.6.0
wsproto==1.2.0
zope.event==5.0
zope.interface==6.3
